rename component/ 로 navigation.tsx 옮김
