+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css'
 import { Route, Routes } from 'react-router-dom';
-import NavigationBar from './components/stylePage/NavigationBar';
+import NavigationBar from './components/NavigationBar';
 import Home from './pages/Home';
 import Content from './pages/Content';
 import Style from './pages/Style';
